Restructure directory structure
The structure of the directories was messed up, now we have a more
modular approach with several *.pro files for better organization.
