+++ client/client.pro
@@ -2,15 +2,14 @@ QT += core
 QT += network
 QT -= gui
 
-TARGET = Qt-Client
+TARGET = client
 CONFIG += console
 CONFIG -= app_bundle
 
 TEMPLATE = app
 
-SOURCES += main.cpp \
-    client.cpp
+SOURCES += src/main.cpp \
+    src/client.cpp
 
 HEADERS += \
-    client.h
-
+    src/client.h
