+++ client/src/main.cpp
@@ -1,5 +1,5 @@
 #include <QCoreApplication>
-#include <client.h>
+#include "client.h"
 
 int main(int argc, char *argv[])
 {
