+++ octopus.pro
@@ -1,0 +1,6 @@
+TEMPLATE = subdirs
+
+SUBDIRS += \
+    client \
+    server \
+    gui
