--- server/Qt-Server.pro
@@ -1,18 +1,0 @@
-QT += core
-QT += network
-QT -= gui
-
-TARGET = Qt-Server
-CONFIG += console
-CONFIG -= app_bundle
-
-TEMPLATE = app
-
-SOURCES += main.cpp \
-    myserver.cpp \
-    mythread.cpp
-
-HEADERS += \
-    myserver.h \
-    mythread.h
-
