--- server_wGUI/QT-Server_wGUI/myserver.h
@@ -1,27 +1,0 @@
-
-#ifndef MYSERVER_H
-#define MYSERVER_H
-
-#include <QTcpServer>
-#include <iostream>
-#include <vector>
-#include <string>
-#include <mythread.h>
-
-class MyServer : public QTcpServer
-{
-    Q_OBJECT
-public:
-    explicit MyServer(QObject *parent = 0);
-    void startServer();
-    QList<QTcpSocket*> connections;
-signals:
-
-public slots:
-
-protected:
-    void incomingConnection(qintptr socketDescriptor);
-
-};
-
-#endif // MYSERVER_H
