--- server_wGUI/QT-Server_wGUI/mythread.cpp
@@ -1,68 +1,0 @@
-#include "mythread.h"
-#include "myserver.h"
-
-MyThread::MyThread(qintptr ID, QObject *parent) :
-    QThread(parent)
-{
-    this->socketDescriptor = ID;
-}
-
-void MyThread::run()
-{
-    // thread starts here
-    qDebug() << " Thread started";
-
-    socket = new QTcpSocket();
-
-    // set the ID
-    if(!socket->setSocketDescriptor(this->socketDescriptor))
-    {
-        // something's wrong, we just emit a signal
-        emit error(socket->error());
-        return;
-    }
-
-    // connect socket and signal
-    // note - Qt::DirectConnection is used because it's multithreaded
-    //        This makes the slot to be invoked immediately, when the signal is emitted.
-
-    connect(socket, SIGNAL(readyRead()), this, SLOT(readyRead()), Qt::DirectConnection);
-    connect(socket, SIGNAL(disconnected()), this, SLOT(disconnected()));
-
-    // We'll have multiple clients, we want to know which is which
-    qDebug() << socketDescriptor<< "        IP:"<< socket->peerAddress().toString()<< " Client connected";
-
-    // make this thread a loop,
-    // thread will stay alive so that signal/slot to function properly
-    // not dropped out in the middle when thread dies
-
-    exec();
-}
-
-void MyThread::readyRead()
-{
-    // get the information
-    QByteArray Data = socket->readAll();
-
-    // will write on server side window
-    qDebug() << socketDescriptor<< "        IP:"<< socket->peerAddress().toString() << "       Data in: " << Data;
-
-    if(Data == "Ready for Command\n"){
-        socket->write("COMMAND!!!");
-        int i = 0;
-        for(i = 0; i<this->parent()->findChildren<MyThread*>().size(); i++ ){
-            qDebug() << this->parent()->findChildren<MyThread*>().at(i)->socketDescriptor << "       "<< this->parent()->findChildren<MyThread*>().at(i)->socket->peerAddress().toString();
-        }
-    }else{
-        socket->write(Data);
-    }
-}
-
-void MyThread::disconnected()
-{
-    qDebug() << socketDescriptor << " Disconnected";
-
-
-    socket->deleteLater();
-    exit(0);
-}
