--- server_wGUI/QT-Server_wGUI/mythread.h
@@ -1,28 +1,0 @@
-#ifndef MYTHREAD_H
-#define MYTHREAD_H
-
-#include <QThread>
-#include <QTcpSocket>
-#include <QDebug>
-
-class MyThread : public QThread
-{
-    Q_OBJECT
-public:
-    explicit MyThread(qintptr ID, QObject *parent = 0);
-
-    void run();
-
-signals:
-    void error(QTcpSocket::SocketError socketerror);
-
-public slots:
-    void readyRead();
-    void disconnected();
-
-public:
-    QTcpSocket *socket;
-    qintptr socketDescriptor;
-};
-
-#endif // MYTHREAD_H
